Tried to profile and optimize memory
