+++ Gemfile
@@ -3,3 +3,5 @@ source "https://rubygems.org"
 
 gem 'minitest'
 gem 'ruby-prof'
+gem 'memory_profiler'
+gem 'pry'
