+++ README.md
@@ -16,7 +16,6 @@
 
 Можем считать, что все сессии юзера всегда идут одним непрерывным куском. Нет такого, что сначала идёт часть сессий юзера, потом сессии другого юзера, и потом снова сессии первого.
 
-
 ## План работы
 В этот раз переработка потребуется кардинальная, так что нужно сделать два этапа
