+++ case-study.md
@@ -45,8 +45,13 @@ puts "MEMORY USAGE: %d MB" % (`ps -o rss= -p #{Process.pid}`.to_i / 1024)
 ```
 
 На этапе отладки потокового варианта использовала sample, к-й позволял оценить работоспособность в пределах пары секунд.
-Далее использовала простой вывод, чтобы оценивать в пределах 35 секунд, либо sample, чтобы делать это быстрее.
 
+Далее подобрала sample, с к-м фидбек можно было получить за n секунд.
+
+
+Далее использовала простой вывод, чтобы оценивать в пределах 35 секунд.
+
+Либо sample на 1_000_000 строк, тогда в пределах 7 секунд.
 
 ## Вникаем в детали системы, чтобы найти главные точки роста
 Для того, чтобы найти "точки роста" для оптимизации я воспользовался *инструментами, которыми вы воспользовались*
@@ -68,6 +73,55 @@ puts "MEMORY USAGE: %d MB" % (`ps -o rss= -p #{Process.pid}`.to_i / 1024)
 Думала использовать `gem 'json-write-stream'` , но неудобно писать вложенный json для `usersStats`, поэтому решила напрямую
 
 ### Ваша находка №1
+
+memory profiler показал ,что File очень много ест памяти, но хз, как от него избавиться!
+
+Ещё String, попробую от неё избавиться. Тут ничего не изменилось.
+
+Также посмотрела qcachegrind (у меня kcachegrind) - много отнимают `write_user_to_json` , `Array#each`.
+
+
+Также создала второй тред, вот его отчёт:
+
+```
+MEMORY USAGE: 29 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+MEMORY USAGE: 334 MB
+22.006778048
+```
+
+
+
+
+Подумала, раз `File`, то не открывать/закрывать его для каждого юзера, а открыть на запись 1 раз.
+
+Переписала - по памяти ничего не поменялось, но программа стала быстрее отрабатывать по рвемени.
+
+
+
+
+
+
+
 - какой отчёт показал главную точку роста
 - как вы решили её оптимизировать
 - как изменилась метрика
